Added Encrypted Update WiFiManager.cpp
This change allows setting a 256bit encryption key and address location of the main code partition to the setup.    If the key is setup, all OTA updates must be encrypted.   This protects the firmware from copying because the decryption key is in the data partition on the ESP32.    This allows arduino apps to go into production simply by blowing the download_mode and jtag fuses.   The firmware will be protected from copying without having to setup the arduino incompatible secure boot.   The normal boot partition from arduino will work fine.